backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
httpx
python-multipart
sqlalchemy
python-dotenv
